packages/Transformer2/requirements.txt
absl-py
nltk
numpy
six>=1.14
rouge-score